Shut up a couple of warnings occuring while compiling on Solaris 10
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@664 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ config.c
@@ -655,7 +655,7 @@ void check_number_parse_errno(char c, const char *oarg, char *endptr) {
 }
 
 static uint16_t tou16(int in, const char *comment) {
-    if(in > UINT16_MAX) {
+    if(in > (int) UINT16_MAX) {
 	fprintf(stderr, "Number of %s %d too big\n", comment, in);
 	exit(1);
     }
